Handle occurrences of '}' within an antiquotation
When encountering an antiquotation, we incrementally process the input
character by character, applying Haskell's string quoting syntax rules
to determine whether an encountered '}' is the terminator of the
antiquoted expression, or if it is a part of the antiquoted expression
itself.

This *should* cover all possible occurrences of '}' in a valid Haskell
98/Haskell 2010 expression, but no liability is assumed by the author in
any overlooked corner cases! Bug reports, however, are quite welcome.