Fix infinite loop with ${}
Prior to this commit, the quasiquotation [i|${}] would cause an infinite loop
and eat all your RAM. It was parsing it as an infinite list of `Lit []`. This
change removes the `try` around `p_anti`, which causes the parser to commit to
`p_anti` as soon as it sees an unescaped '$' character. When `p_anti` fails, the
whole parse fails.